+++ r-kvstore-20150101/core/Models/DescribeBackupsRequest.cs
@@ -14,11 +14,11 @@ namespace AlibabaCloud.SDK.R_kvstore20150101.Models
         /// </summary>
         [NameInMap("BackupId")]
         [Validation(Required=false)]
-        public int? BackupId { get; set; }
+        public long? BackupId { get; set; }
 
         [NameInMap("BackupJobId")]
         [Validation(Required=false)]
-        public int? BackupJobId { get; set; }
+        public long? BackupJobId { get; set; }
 
         /// <summary>
         /// The end of the time range to query. Specify the time in the *yyyy-MM-dd*T*HH:mm*Z format. The time must be in UTC. The end time must be later than the start time.
